+++ src/SolidObject.js
@@ -53,11 +53,13 @@ function guid () {
  */
 const Symbols = {
   state: Symbol( '__state__' ),
-  props: Symbol(' __props__' ),
+  props: Symbol('__props__' ),
   key: Symbol( '__key__' ),
   defaults: Symbol( '__defaults__' ),
   solid: Symbol( '__solid__' )
 }
+
+const merge = require( './merge' )
 
 /**
  * @name module:SolidObject~SolidObject
@@ -144,7 +146,7 @@ class SolidObject {
 
     if ( this.propsShouldUpdate() ) {
       props = this.propsWillUpdate( props )
-      this[ Symbols.props ] = Object.assign( this.props,  props )
+      this[ Symbols.props ] = merge( this.props,  props )
       this.propsDidUpdate()
     }