+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.1] 2016-09-09
+
+### CHANGED
+* SolidObject `setProps` will not deep-merge the new properties.
+
 ## [1.0.0] 2016-08-29
 
 ### ADDED
-
-  * SolidObject class.
+* SolidObject class.